pyproject.toml
[tool.poetry]
name = "epidemiologic-calculator"
version = "0.1.0"
description = "It's a backend API to calculate epidemic parameters, and make projections."
authors = ["DataScienceResearchPeru"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
bcrypt = "^3.1.7"
Flask = "^1.1.2"
Flask-Cors = "^3.0.8"
Flask-Injector = "^0.12.3"
Flask-JWT-Extended = "^3.24.1"
Flask-Mail = "^0.9.1"
Flask-RESTful = "^0.3.8"
flask-restplus = "^0.13.0"
Flask-SQLAlchemy = "^2.4.1"
gunicorn = "^20.0.4"
injector = "^0.18.3"
Jinja2 = "^2.11.1"
numpy = "^1.18.2"
psycopg2-binary = "^2.8.5"
PyJWT = "^1.7.1"
python-dotenv = "^0.12.0"
scipy = "^1.4.1"
SQLAlchemy = "^1.3.16"
click = "^7.1.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.1"
bandit = "^1.6.2"
black = "^19.10b0"
codecov = "^2.0.22"
coverage = "^5.1"
flake8 = "^3.7.9"
flake8-black = "^0.1.1"
mypy = "^0.770"
nb_black = "^1.0.7"
nbstripout = "^0.3.7"
pre-commit = "^2.2.0"
pycodestyle = "^2.5.0"
pydocstyle = "^5.0.2"
pylint = "^2.4.4"
pylint-flask = "^0.6"
pylint-plugin-utils = "^0.6"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-mock = "^3.0.0"
pytest-vcr = "^1.0.2"
pytest-xdist = "^1.31.0"
sqlalchemy-stubs = "^0.3"
tox = "^3.14.6"

[tool.black]
target_version = ['py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |\.hg
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |_build
  |buck-out
  |build
  |dist
  |templates
  |locale
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
